adicionados metodos de gravar endereços e telefone do usuario

+++ src/main/java/com/JavaSpring/usuario/dto/EnderecoDTO.java
@@ -12,7 +12,7 @@ public class EnderecoDTO {
 
     private String rua;
     private String cidade;
-    private long numero;
+    private Double numero;
     private String complemento;
     private String cep;
 }
